add abac

+++ app/routers/files.py
@@ -1,5 +1,5 @@
 from fastapi import Depends, FastAPI, HTTPException, APIRouter, UploadFile
-from ..internal.auth import get_current_user
+from ..internal.auth import get_current_user, verify_access
 import os
 import shutil
 
@@ -11,11 +11,12 @@ router = APIRouter(
 )
 
 
-UPLOAD_DIRECTORY = "./deploy"
+UPLOAD_DIRECTORY = "deploy"
 
 
 @router.post("/")
 async def create_package(file: UploadFile):
+    verify_access(1)
     file_location = os.path.join(UPLOAD_DIRECTORY, file.filename)
     if os.path.isfile(file_location):
         return HTTPException(status_code=400, detail="file name already exists")
@@ -25,6 +26,7 @@ async def create_package(file: UploadFile):
 
 @router.delete("/delete/")
 def delete_file(filename: str):
+    verify_access(1)
     file_location = os.path.join(UPLOAD_DIRECTORY, filename)
     if os.path.exists(file_location):
         os.remove(file_location)
